Markers now filtering away.
Finally discovered bug in updating markers: the directive doesn't detect a change when if the 'key' used to reference the marker is the same.  So if you use an array, the 'keys' (in this case the id of item in the array) are the same, as they are if you generate m0, m1 etc starting from 0 each time.  The solution was to generate a count that only goes up and never resets to 0.