Add ability to add resource from map view
If not logged in, shows login form.  Once logged in, presents form to fill out. If not registered, clicking register will show register form.  On registration, user will be prompted to login with their username populated.  Also added ability to press and hold to save a location on the map.